feat product controller

+++ database/migrations/2023_11_12_072200_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nama');
             $table->string('desc');
-            $table->double('size');
+            $table->integer('size');
             $table->string('warna');
             $table->double('berat');
             $table->integer('stok');
